+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-oracle/pom.xml
@@ -26,7 +26,11 @@
         <version>${revision}</version>
     </parent>
     <artifactId>connector-cdc-oracle</artifactId>
-    <name>SeaTunnel : Connectors V2 : CDC : oracle</name>
+    <name>SeaTunnel : Connectors V2 : CDC : Oracle</name>
+
+    <properties>
+        <oracle.version>19.18.0.0</oracle.version>
+    </properties>
 
     <dependencyManagement>
         <dependencies>
@@ -48,10 +52,20 @@
                 <groupId>org.apache.seatunnel</groupId>
                 <artifactId>connector-jdbc</artifactId>
                 <version>${project.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
             </dependency>
 
+            <dependency>
+                <groupId>com.oracle.database.jdbc</groupId>
+                <artifactId>ojdbc8</artifactId>
+                <version>${oracle.version}</version>
+                <scope>provided</scope>
+            </dependency>
+            <dependency>
+                <groupId>com.oracle.database.xml</groupId>
+                <artifactId>xdb</artifactId>
+                <version>${oracle.version}</version>
+                <scope>provided</scope>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
@@ -70,39 +84,16 @@
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
             <artifactId>connector-jdbc</artifactId>
-            <version>${project.version}</version>
         </dependency>
 
         <dependency>
             <groupId>com.oracle.database.jdbc</groupId>
             <artifactId>ojdbc8</artifactId>
         </dependency>
-
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.13.2</version>
-            <scope>test</scope>
+            <groupId>com.oracle.database.xml</groupId>
+            <artifactId>xdb</artifactId>
         </dependency>
     </dependencies>
 
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>2.4</version>
-                <configuration>
-                    <skip>false</skip>
-                </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>test-jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
 </project>
