+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-oracle/src/main/java/org/apache/seatunnel/connectors/seatunnel/cdc/oracle/source/reader/fetch/scan/OracleSnapshotSplitReadTask.java
@@ -1,0 +1,281 @@
+package org.apache.seatunnel.connectors.seatunnel.cdc.oracle.source.reader.fetch.scan;
+
+import org.apache.seatunnel.connectors.cdc.base.relational.JdbcSourceEventDispatcher;
+import org.apache.seatunnel.connectors.cdc.base.source.split.SnapshotSplit;
+import org.apache.seatunnel.connectors.cdc.base.source.split.wartermark.WatermarkKind;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.source.offset.RedoLogOffset;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleConnectionUtils;
+import org.apache.seatunnel.connectors.seatunnel.cdc.oracle.utils.OracleUtils;
+
+import org.apache.kafka.connect.data.Field;
+import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.data.SchemaBuilder;
+import org.apache.kafka.connect.errors.ConnectException;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import io.debezium.DebeziumException;
+import io.debezium.connector.oracle.OracleConnection;
+import io.debezium.connector.oracle.OracleConnectorConfig;
+import io.debezium.connector.oracle.OracleDatabaseSchema;
+import io.debezium.connector.oracle.OracleOffsetContext;
+import io.debezium.connector.oracle.OracleValueConverters;
+import io.debezium.pipeline.EventDispatcher;
+import io.debezium.pipeline.source.AbstractSnapshotChangeEventSource;
+import io.debezium.pipeline.source.spi.SnapshotProgressListener;
+import io.debezium.pipeline.spi.ChangeRecordEmitter;
+import io.debezium.pipeline.spi.OffsetContext;
+import io.debezium.pipeline.spi.SnapshotResult;
+import io.debezium.relational.Column;
+import io.debezium.relational.RelationalSnapshotChangeEventSource;
+import io.debezium.relational.SnapshotChangeRecordEmitter;
+import io.debezium.relational.Table;
+import io.debezium.relational.TableId;
+import io.debezium.relational.ValueConverter;
+import io.debezium.util.Clock;
+import io.debezium.util.ColumnUtils;
+import io.debezium.util.Strings;
+import io.debezium.util.Threads;
+
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.time.Duration;
+
+/** A wrapped task to fetch snapshot split of table. */
+public class OracleSnapshotSplitReadTask extends AbstractSnapshotChangeEventSource {
+
+    private static final Logger LOG = LoggerFactory.getLogger(OracleSnapshotSplitReadTask.class);
+
+    /** Interval for showing a log statement with the progress while scanning a single table. */
+    private static final Duration LOG_INTERVAL = Duration.ofMillis(10_000);
+
+    private final OracleConnectorConfig connectorConfig;
+    private final OracleDatabaseSchema databaseSchema;
+    private final OracleConnection jdbcConnection;
+    private final JdbcSourceEventDispatcher dispatcher;
+    private final Clock clock;
+    private final SnapshotSplit snapshotSplit;
+    private final OracleOffsetContext offsetContext;
+    private final SnapshotProgressListener snapshotProgressListener;
+
+    public OracleSnapshotSplitReadTask(
+            OracleConnectorConfig connectorConfig,
+            OracleOffsetContext previousOffset,
+            SnapshotProgressListener snapshotProgressListener,
+            OracleDatabaseSchema databaseSchema,
+            OracleConnection jdbcConnection,
+            JdbcSourceEventDispatcher dispatcher,
+            SnapshotSplit snapshotSplit) {
+        super(connectorConfig, snapshotProgressListener);
+        this.offsetContext = previousOffset;
+        this.connectorConfig = connectorConfig;
+        this.databaseSchema = databaseSchema;
+        this.jdbcConnection = jdbcConnection;
+        this.dispatcher = dispatcher;
+        this.clock = Clock.SYSTEM;
+        this.snapshotSplit = snapshotSplit;
+        this.snapshotProgressListener = snapshotProgressListener;
+    }
+
+    @Override
+    public SnapshotResult execute(ChangeEventSourceContext context, OffsetContext previousOffset)
+            throws InterruptedException {
+        SnapshottingTask snapshottingTask = getSnapshottingTask(previousOffset);
+        final SnapshotContext ctx;
+        try {
+            ctx = prepare(context);
+        } catch (Exception e) {
+            LOG.error("Failed to initialize snapshot context.", e);
+            throw new RuntimeException(e);
+        }
+        try {
+            return doExecute(context, previousOffset, ctx, snapshottingTask);
+        } catch (InterruptedException e) {
+            LOG.warn("Snapshot was interrupted before completion");
+            throw e;
+        } catch (Exception t) {
+            throw new DebeziumException(t);
+        } finally {
+            complete(ctx);
+        }
+    }
+
+    @Override
+    protected SnapshotResult doExecute(
+            ChangeEventSourceContext context,
+            OffsetContext previousOffset,
+            SnapshotContext snapshotContext,
+            SnapshottingTask snapshottingTask)
+            throws Exception {
+        final RelationalSnapshotChangeEventSource.RelationalSnapshotContext ctx =
+                (RelationalSnapshotChangeEventSource.RelationalSnapshotContext) snapshotContext;
+        ctx.offset = offsetContext;
+
+        final RedoLogOffset lowWatermark =
+                OracleConnectionUtils.currentRedoLogOffset(jdbcConnection);
+        LOG.info(
+                "Snapshot step 1 - Determining low watermark {} for split {}",
+                lowWatermark,
+                snapshotSplit);
+        ((SnapshotSplitChangeEventSourceContext) context).setLowWatermark(lowWatermark);
+        dispatcher.dispatchWatermarkEvent(
+                offsetContext.getPartition(), snapshotSplit, lowWatermark, WatermarkKind.LOW);
+
+        LOG.info("Snapshot step 2 - Snapshotting data");
+        createDataEvents(ctx, snapshotSplit.getTableId());
+
+        final RedoLogOffset highWatermark =
+                OracleConnectionUtils.currentRedoLogOffset(jdbcConnection);
+        LOG.info(
+                "Snapshot step 3 - Determining high watermark {} for split {}",
+                highWatermark,
+                snapshotSplit);
+        ((SnapshotSplitChangeEventSourceContext) context).setHighWatermark(highWatermark);
+        dispatcher.dispatchWatermarkEvent(
+                offsetContext.getPartition(), snapshotSplit, highWatermark, WatermarkKind.HIGH);
+        return SnapshotResult.completed(ctx.offset);
+    }
+
+    @Override
+    protected SnapshottingTask getSnapshottingTask(OffsetContext previousOffset) {
+        return new SnapshottingTask(false, true);
+    }
+
+    @Override
+    protected SnapshotContext prepare(ChangeEventSourceContext changeEventSourceContext)
+            throws Exception {
+        if (connectorConfig.getPdbName() != null) {
+            jdbcConnection.setSessionToPdb(connectorConfig.getPdbName());
+        }
+        return new OracleSnapshotContext();
+    }
+
+    @Override
+    protected void complete(SnapshotContext snapshotContext) {
+        if (connectorConfig.getPdbName() != null) {
+            jdbcConnection.resetSessionToCdb();
+        }
+    }
+
+    private void createDataEvents(
+            RelationalSnapshotChangeEventSource.RelationalSnapshotContext snapshotContext,
+            TableId tableId)
+            throws Exception {
+        EventDispatcher.SnapshotReceiver snapshotReceiver =
+                dispatcher.getSnapshotChangeEventReceiver();
+        LOG.debug("Snapshotting table {}", tableId);
+        createDataEventsForTable(
+                snapshotContext, snapshotReceiver, databaseSchema.tableFor(tableId));
+        snapshotReceiver.completeSnapshot();
+    }
+
+    /** Dispatches the data change events for the records of a single table. */
+    private void createDataEventsForTable(
+            RelationalSnapshotChangeEventSource.RelationalSnapshotContext snapshotContext,
+            EventDispatcher.SnapshotReceiver snapshotReceiver,
+            Table table)
+            throws InterruptedException {
+
+        long exportStart = clock.currentTimeInMillis();
+        LOG.info("Exporting data from split '{}' of table {}", snapshotSplit.splitId(), table.id());
+
+        final String selectSql =
+                OracleUtils.buildSplitScanQuery(
+                        snapshotSplit.getTableId(),
+                        snapshotSplit.getSplitKeyType(),
+                        snapshotSplit.getSplitStart() == null,
+                        snapshotSplit.getSplitEnd() == null);
+        LOG.info(
+                "For split '{}' of table {} using select statement: '{}'",
+                snapshotSplit.splitId(),
+                table.id(),
+                selectSql);
+
+        try (PreparedStatement selectStatement =
+                        OracleUtils.readTableSplitDataStatement(
+                                jdbcConnection,
+                                selectSql,
+                                snapshotSplit.getSplitStart() == null,
+                                snapshotSplit.getSplitEnd() == null,
+                                snapshotSplit.getSplitStart(),
+                                snapshotSplit.getSplitEnd(),
+                                snapshotSplit.getSplitKeyType().getTotalFields(),
+                                connectorConfig.getSnapshotFetchSize());
+                ResultSet rs = selectStatement.executeQuery()) {
+
+            ColumnUtils.ColumnArray columnArray = ColumnUtils.toArray(rs, table);
+            long rows = 0;
+            Threads.Timer logTimer = getTableScanLogTimer();
+
+            while (rs.next()) {
+                rows++;
+                final Object[] row = new Object[columnArray.getGreatestColumnPosition()];
+                for (int i = 0; i < columnArray.getColumns().length; i++) {
+                    Column actualColumn = table.columns().get(i);
+                    row[columnArray.getColumns()[i].position() - 1] =
+                            readField(rs, i + 1, actualColumn, table);
+                }
+                if (logTimer.expired()) {
+                    long stop = clock.currentTimeInMillis();
+                    LOG.info(
+                            "Exported {} records for split '{}' after {}",
+                            rows,
+                            snapshotSplit.splitId(),
+                            Strings.duration(stop - exportStart));
+                    snapshotProgressListener.rowsScanned(table.id(), rows);
+                    logTimer = getTableScanLogTimer();
+                }
+                dispatcher.dispatchSnapshotEvent(
+                        table.id(),
+                        getChangeRecordEmitter(snapshotContext, table.id(), row),
+                        snapshotReceiver);
+            }
+            LOG.info(
+                    "Finished exporting {} records for split '{}', total duration '{}'",
+                    rows,
+                    snapshotSplit.splitId(),
+                    Strings.duration(clock.currentTimeInMillis() - exportStart));
+        } catch (SQLException e) {
+            throw new ConnectException("Snapshotting of table " + table.id() + " failed", e);
+        }
+    }
+
+    protected ChangeRecordEmitter getChangeRecordEmitter(
+            SnapshotContext snapshotContext, TableId tableId, Object[] row) {
+        snapshotContext.offset.event(tableId, clock.currentTime());
+        return new SnapshotChangeRecordEmitter(snapshotContext.offset, row, clock);
+    }
+
+    private Threads.Timer getTableScanLogTimer() {
+        return Threads.timer(clock, LOG_INTERVAL);
+    }
+
+    /**
+     * copied from io.debezium.connector.oracle.antlr.listener.ParserUtils#convertValueToSchemaType.
+     */
+    private Object readField(ResultSet rs, int fieldNo, Column actualColumn, Table actualTable)
+            throws SQLException {
+
+        OracleValueConverters oracleValueConverters =
+                new OracleValueConverters(connectorConfig, jdbcConnection);
+
+        final SchemaBuilder schemaBuilder = oracleValueConverters.schemaBuilder(actualColumn);
+        if (schemaBuilder == null) {
+            return null;
+        }
+        Schema schema = schemaBuilder.build();
+        Field field = new Field(actualColumn.name(), 1, schema);
+        final ValueConverter valueConverter = oracleValueConverters.converter(actualColumn, field);
+        return valueConverter.convert(rs.getObject(fieldNo));
+    }
+
+    private static class OracleSnapshotContext
+            extends RelationalSnapshotChangeEventSource.RelationalSnapshotContext {
+
+        public OracleSnapshotContext() throws SQLException {
+            super("");
+        }
+    }
+}
