--- seatunnel-connectors-v2/connector-cdc/connector-cdc-oracle/src/main/java/org/apache/seatunnel/connectors/seatunnel/cdc/oracle/source/source/OracleSourceOptions.java
@@ -1,68 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.seatunnel.connectors.seatunnel.cdc.oracle.source.source;
-
-import org.apache.seatunnel.api.configuration.Option;
-import org.apache.seatunnel.api.configuration.Options;
-import org.apache.seatunnel.api.configuration.SingleChoiceOption;
-import org.apache.seatunnel.connectors.cdc.base.option.SourceOptions;
-import org.apache.seatunnel.connectors.cdc.base.option.StartupMode;
-import org.apache.seatunnel.connectors.cdc.base.option.StopMode;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class OracleSourceOptions {
-    public static final SingleChoiceOption<StartupMode> STARTUP_MODE =
-            (SingleChoiceOption)
-                    Options.key(SourceOptions.STARTUP_MODE_KEY)
-                            .singleChoice(
-                                    StartupMode.class,
-                                    Arrays.asList(
-                                            StartupMode.INITIAL,
-                                            StartupMode.EARLIEST,
-                                            StartupMode.LATEST))
-                            .defaultValue(StartupMode.INITIAL)
-                            .withDescription(
-                                    "Optional startup mode for CDC source, valid enumerations are "
-                                            + "\"initial\", \"earliest\", \"latest\", \"timestamp\"\n or \"specific\"");
-
-    public static final SingleChoiceOption<StopMode> STOP_MODE =
-            (SingleChoiceOption)
-                    Options.key(SourceOptions.STOP_MODE_KEY)
-                            .singleChoice(StopMode.class, Arrays.asList(StopMode.NEVER))
-                            .defaultValue(StopMode.NEVER)
-                            .withDescription(
-                                    "Optional stop mode for CDC source, valid enumerations are "
-                                            + "\"never\", \"latest\", \"timestamp\"\n or \"specific\"");
-
-    public static final Option<List<String>> SCHEMA_NAMES =
-            Options.key("schema-names")
-                    .listType()
-                    .noDefaultValue()
-                    .withDescription("Schema name of the database to monitor.");
-
-    public static final Option<String> SCAN_INCREMENTAL_SNAPSHOT_CHUNK_KEY_COLUMN =
-            Options.key("scan.incremental.snapshot.chunk.key-column")
-                    .stringType()
-                    .noDefaultValue()
-                    .withDescription(
-                            "The chunk key of table snapshot, captured tables are split into multiple chunks by a chunk key when read the snapshot of table."
-                                    + "By default, the chunk key is the first column of the primary key and the chunk key is the RowId in oracle."
-                                    + "This column must be a column of the primary key.");
-}
