+++ seatunnel-connectors-v2/connector-jdbc/src/main/java/org/apache/seatunnel/connectors/seatunnel/jdbc/catalog/oracle/OracleDataTypeConvertor.java
@@ -98,6 +98,9 @@ public class OracleDataTypeConvertor implements DataTypeConvertor<String> {
                         MapUtils.getInteger(dataTypeProperties, PRECISION, DEFAULT_PRECISION);
                 int scale = MapUtils.getInteger(dataTypeProperties, SCALE, DEFAULT_SCALE);
                 if (scale == 0) {
+                    if (precision == 0) {
+                        return new DecimalType(38, 18);
+                    }
                     if (precision == 1) {
                         return BasicType.BOOLEAN_TYPE;
                     }
