+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-cdc-oracle-e2e/src/test/resources/oraclecdc_to_oracle_with_multi_table_mode_two_table.conf
@@ -33,7 +33,7 @@ source {
     password = "oracle"
     database-names = ["XE"]
     schema-names = ["DEBEZIUM"]
-    table-names = ["XE.DEBEZIUM.FULL_TYPES"]
+    table-names = ["XE.DEBEZIUM.FULL_TYPES","XE.DEBEZIUM.FULL_TYPES2"]
     base-url = "jdbc:oracle:thin:system/oracle@oracle-host:1521:xe"
     source.reader.close.timeout = 120000
     debezium {
@@ -48,19 +48,16 @@ transform {
 }
 
 sink {
-  console {
-    source_table_name = "customers"
-  }
 Jdbc {
   source_table_name = "customers"
   driver = "oracle.jdbc.driver.OracleDriver"
   url = "jdbc:oracle:thin:system/oracle@oracle-host:1521:xe"
-  username = "system"
+  user = "system"
   password = "oracle"
   generate_sink_sql = true
+  tablePrefix = "SINK_"
   database = "XE"
-  table = "DEBEZIUM.FULL_TYPES_SINK"
+  schema = "DEBEZIUM"
   batch_size = 1
-  primary_keys = ["ID"]
 }
 }
